Convert values to NA

This is a replacement for dplyr::na_if(). It is useful if you want to convert annoying values to NA. Unlike dplyr::na_if(), this function allows you to specify multiple values to be replaced with NA at the same time.

na_if_in() replaces values that match its arguments with NA.

na_if_not() replaces values that do not match its arguments with NA.

na_if_in(x, ...)

na_if(x, ...)

na_if_not(x, ...)

Arguments

x
Vector to modify

...
Values to replace with NA, specified as either:

An object, vector of objects, or list of objects

A one-sided formula (see section "Formulas")

Value

A modified version of x with selected values replaced with NA.

Formulas

These functions accept one-sided formulas that can evaluate to logical vectors. The input is represented in these conditional statements as ".". Valid formulas take the form ~ . < 0. See examples.

Lifecycle

na_if() has been deprecated in favor of na_if_in() to avoid masking dplyr::na_if().

See also

Examples

-1:10

#>  [1] -1  0  1  2  3  4  5  6  7  8  9 10

# We can replace -1...
# ... explicitly
na_if_in(-1:10, -1)

#>  [1] NA  0  1  2  3  4  5  6  7  8  9 10

# ... by specifying values to keep
na_if_not(-1:10, 0:10)

#>  [1] NA  0  1  2  3  4  5  6  7  8  9 10

# ... using a formula
na_if_in(-1:10, ~ . < 0)

#>  [1] NA  0  1  2  3  4  5  6  7  8  9 10

# ... or using a function
na_if_in(-1:10, min)

#>  [1] NA  0  1  2  3  4  5  6  7  8  9 10

messy_string <- c("abc", "", "def", "NA", "ghi", 42, "jkl", "NULL", "mno")
# We can replace unwanted values...
# ... one at a time
na_if_in(messy_string, "")

#> [1] "abc"  NA     "def"  "NA"   "ghi"  "42"   "jkl"  "NULL" "mno" 

# ... or all at once
na_if_in(messy_string, "", "NA", "NULL", 1:100)

#> [1] "abc" NA    "def" NA    "ghi" NA    "jkl" NA    "mno"

na_if_in(messy_string, c("", "NA", "NULL", 1:100))

#> [1] "abc" NA    "def" NA    "ghi" NA    "jkl" NA    "mno"

na_if_in(messy_string, list("", "NA", "NULL", 1:100))

#> [1] "abc" NA    "def" NA    "ghi" NA    "jkl" NA    "mno"

# ... or using a clever formula
grepl("[a-z]{3,}", messy_string)

#> [1]  TRUE FALSE  TRUE FALSE  TRUE FALSE  TRUE FALSE  TRUE

na_if_not(messy_string, ~ grepl("[a-z]{3,}", .))

#> [1] "abc" NA    "def" NA    "ghi" NA    "jkl" NA    "mno"

# na_if_in() is particularly useful inside dplyr::mutate
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:fauxnaif’:
#> 
#>     na_if

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

faux_census %>%
  mutate(
    state = na_if_in(state, "Canada"),
    age   = na_if_in(age, ~ . < 18, ~ . > 120)
  )

#> # A tibble: 20 x 6
#>    state gender               age race             income religion              
#>    <chr> <chr>              <dbl> <chr>             <dbl> <chr>                 
#>  1 CA    female                80 Native American  2.80e4 Christian             
#>  2 NY    Woman                 89 Latino           1.49e5 Spiritual not religio…
#>  3 CA    Female                48 White            4.79e5 Catholic              
#>  4 TX    Male                  63 latinx           8.50e4 christian             
#>  5 PA    Male                  47 asian            4.19e4 Baptist               
#>  6 TX    Gender is a socia…    57 Race is a socia… 1.00e7 Religion is the opiat…
#>  7 NA    Male                  49 white            1.49e5 methodist             
#>  8 TX    Female                50 White            9.88e4 Lutheran              
#>  9 NY    f                     NA white            9.07e4 Agnostic              
#> 10 WA    F                     33 White            4.50e4 Jewish                
#> 11 TX    Male                  30 White            1.27e5 none                  
#> 12 OH    Non-binary            42 Caucasian        2.16e4 Roman Catholic        
#> 13 NC    Female                22 African American 7.42e4 atheist               
#> 14 LA    Male                  NA White            6.10e4 Christian             
#> 15 LA    Female                28 Black            2.00e4 Not religious         
#> 16 CA    male                  34 Asian American   7.74e4 Christian             
#> 17 TN    M                     64 white            1.00e7 Nothing               
#> 18 FL    Female                68 white            4.71e4 None                  
#> 19 OH    Male                  39 black            2.38e4 baptist               
#> 20 NH    male                  73 Hispanic         3.32e4 Christian             

# We get a message if our values to replace don't exist
na_if_in(-1:10, 11)

#>  [1] -1  0  1  2  3  4  5  6  7  8  9 10

# And a warning if we use an invalid input...
# ... like a two-sided formula
na_if_in(-1:10, x ~ . < 0)

#> Warning: Formulas must be one-sided.
#> * `x ~ . < 0` is a two-sided function.

#>  [1] -1  0  1  2  3  4  5  6  7  8  9 10

# ... NULL
na_if_in(-1:10, NULL)

#> Warning: Arguments must be atomics, formulas or functions. 
#> * `NULL` can't be used.

#>  [1] -1  0  1  2  3  4  5  6  7  8  9 10

# ... or nothing at all
na_if_in(-1:10)

#> Warning: No values to replace with `NA` specified.

#>  [1] -1  0  1  2  3  4  5  6  7  8  9 10

# This function handles vector values differently than dplyr,
# and returns a different result with vector replacement values:
na_if_in(1:5, 5:1)

#> [1] NA NA NA NA NA

dplyr::na_if(1:5, 5:1)

#> [1]  1  2 NA  4  5