Convert values to NA in multiple columns

na_if_all(.tbl, ...)

na_if_not_all(.tbl, ...)

na_if_at(.tbl, .vars, ...)

na_if_not_at(.tbl, .vars, ...)

na_if_if(.tbl, .predicate, ...)

na_if_not_if(.tbl, .predicate, ...)

Arguments

.tbl
A tbl object

...
Values to replace with NA, specified as either:

An object, vector of objects, or list of objects

A one-sided formula (see section "Formulas" in na_if())

.vars
A list of columns generated by dplyr::vars(), a character vector of column names, a numeric vector of column positions, or NULL.

.predicate
A predicate function to be applied to the columns or a logical vector. The variables for which .predicate is or returns TRUE are selected. This argument is passed to rlang::as_function() and thus supports quosure-style lambda functions and strings representing function names.

Value

A modified data frame. Matched values in selected columns are replaced with NA.

Details

The dplyr::scoped variants of na_if() and na_if_not() can be used directly within pipelines and can modify multiple variables at once.

*_all() affects every variable

*_at() affects variables selected with a character vector or dplyr::vars()

*_if() affects variables selected with a predicate function

See also

Examples

if (FALSE) {
df <- data.frame(a = 0:5, b = 5:0, c = as.numeric(0:5), d = letters[1:6])

na_if_all(df, 0)
na_if_not_all(df, 0:3, "c")

na_if_at(df, c("a", "c"), 0)
na_if_not_at(df, c("a", "c"), 0:3)

na_if_if(df, is.integer, 0)
na_if_not_if(df, is.integer, 0:3)
}